requirements.txt
fastapi
uvicorn
python-multipart
pillow
numpy
imagehash
scipy
opencv-python-headless
pytesseract